Use GNOME Shell Introspection API
`Eval` was recently disabled unless the unsafe context was enabled:
https://gitlab.gnome.org/GNOME/gnome-shell/-/commit/7298ee23e91b756c7009b4d7687dfd8673856f8b
Since all we need is to get the current focused window, we can avoid the
unsafe context requirement by using the Introspection API instead.

I have also updated zbus in order to get the Variant API's to help with
deserialization.